Moved walk_config_signals to _readable
